fix: valid STOP completion results in an error being returned  (#875)
- There's currently a
[bug](https://github.com/google/adk-python/blob/14484065c64396cebc4a1dde84d6b8b51439b990/src/google/adk/models/llm_response.py#L133C1-L134C1)
in ADK where a valid finish reason ie: `STOP` is passed as the
`error_code` in the llm response when there is no content. This results
in kagent-adk picking up the valid finish reason as an error and emits a
generic `An error occurred during processing` message to the user.
- This PR fixes this issue in kagent-adk and improves the error
messaging by only emitting an error status event for completion reasons
that represent actual failures.

---------

Signed-off-by: JM Huibonhoa <jm.huibonhoa@solo.io>