(Feat) Implement `init`, `build` and `deploy` cmd for kagent cli (#838)
This PR adds the following new features to the `kagent` CLI.

1. `kagent init` creates a bootstrap agent project. 
- Currently supports bootstrapping an agent project for the following
model providers. The supported models are the models that we currently
support in
[models.go](https://github.com/kagent-dev/kagent/blob/main/go/internal/httpserver/handlers/models.go)
       1. OpenAI
       2. Anthropic
       3. AzureOpenAI
       4. Gemini
2. `kagent build` builds a docker image for an agent project.
- By default the image assumes the repository is `localhost:5001`, the
name of the image is the `name` specified in the agent project's
manifest and the tag is `latest`.
3. `kagent deploy` deploys the agent to a kubernetes environement by
creating a .

Tested on the following provider and model combinations.
1. Gemini:  gemini-2.0-flash ✅ 
2. OpenAI: gpt-4 ✅ 

Following provider and model combinations are currently being tested
1. AzureOpenAI: gpt-4 (in progress)
2. Anthropic: claude-sonnet-4 (in progress)

### Follow Ups
1. Add support for `VertexAI` and `AnthropicVertexAI`. I'm currently
blocked on procuring working credentials which is why I have temporarily
removed some of my previous implementation and am opting to add support
for the above in a follow up PR.

---------

Signed-off-by: JM Huibonhoa <jm.huibonhoa@solo.io>