bugfix: fix comparison of agent ids (#957)
Agent id `agent/ns` that is stored in a session is represented as
`agent__NS__ns`. Whereas the same id is represented as `agent/ns` when
it comes from the Principal. A check comparing expected agent id vs.
current thus always fails.

This PR fixes the issue by normalizing the agent id format before the
check to the one used when persisting the id in a db.

Signed-off-by: Dmitri Dolguikh <dmitri@appliedlogic.ca>